pyproject.toml
[project]
name = "school-guardian"
version = "0.1.0"
description = "OpenClaw school-guardian agent for Classroom sync and school task tracking."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.0",
    "sqlalchemy>=2.0.40",
]

[project.optional-dependencies]
google = [
    "google-api-python-client>=2.170.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
]
extract = [
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
]
export = [
    "markdown>=3.8",
    "fpdf2>=2.8.0",
]

[tool.uv]
package = false
